refactor: reservation-schedule-theme 등록/삭제 조건 추가

+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,5 +1,7 @@
 package nextstep.schedule;
 
+import nextstep.reservation.ReservationDao;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -8,16 +10,21 @@ import java.util.List;
 
 @Service
 public class ScheduleService {
-    private ScheduleDao scheduleDao;
-    private ThemeDao themeDao;
+    private final ScheduleDao scheduleDao;
+    private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
+    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao, ReservationDao reservationDao) {
         this.scheduleDao = scheduleDao;
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public Long create(ScheduleRequest scheduleRequest) {
         Theme theme = themeDao.findById(scheduleRequest.getThemeId());
+        if (theme == null) {
+            throw new NullPointerException("테마를 찾을 수 없음");
+        }
         return scheduleDao.save(scheduleRequest.toEntity(theme));
     }
 
@@ -26,6 +33,9 @@ public class ScheduleService {
     }
 
     public void deleteById(Long id) {
+        if (reservationDao.findByScheduleId(id) != null) {
+            throw new NotExistEntityException("예약이 존재하여 삭제할 수 없음");
+        }
         scheduleDao.deleteById(id);
     }
 }
